Draft xcore structure done
