+++ plugins/org.preesm.model.scenario/model/Scenario.xcore
@@ -547,11 +547,43 @@ class PapiCpuID {
  ##################### */
  
  class EnergyConfig{
+ 	//TODO: export/import energy
+	String excelFileURL // excel or csv local file
 	refers Scenario scenario opposite energyConfig
  	contains PerformanceObjective performanceObjective
+ 	//TODO: fill this with real values in the UI and parse/write
+ 	contains PlatformPower platformPower
+ 	//TODO: fill this with real values in the UI and parse/write
+ 	contains PeActorsEnergy[] algorithmEnergy
+ 	//TODO: fill this with real values in the UI and parse/write
+ 	contains PeCommsEnergy[] commsEnergy
  }
  class PerformanceObjective{
  	float objectiveEPS
  	float toleranceEPS
  }
+class PlatformPower{
+	double basePower
+	contains PEPower[] pePower
+}
+class PEPower wraps java.util.Map$Entry {
+	String key 		// The component type
+	Double value	// The power value
+}
+class PeActorsEnergy wraps java.util.Map$Entry {
+	String key 						// The component type
+	contains ActorEnergy[] value	// The actor energy values for this type of PE
+}
+class ActorEnergy wraps java.util.Map$Entry {
+	refers AbstractActor key 	// The actor
+	Double value				// The energy value
+}
+class PeCommsEnergy wraps java.util.Map$Entry{
+	String key					// The source component type
+	contains NodeEnergy[] value // The energy value for each destination
+}
+class NodeEnergy wraps java.util.Map$Entry {
+	String key 		// The destination component type
+	Double value	// The energy value
+}
