+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/editor/messages.properties
@@ -173,9 +173,12 @@ Energy.actorsDescription=This section includes all the information related to en
 Energy.objectiveTitle=Performance objective
 Energy.objectiveDescription=This section contains the objective to reach when performing energy-aware mapping/scheduling
 Energy.platformTitle=Platform power consumption
-Energy.platformDescription=This section includes all the information related to power consumed by the platform resources (static power consumption)
+Energy.platformDescription=This section includes all the information related to power consumed by the platform resources. That is, the static power consumption of the platform
 Energy.commsTitle=Communication energy consumption
 Energy.commsDescription=This section includes all the information related to the energy consumption due to data communication
 Energy.fileRefresh=Refresh
 Energy.energyExportExcel=Export
-
+Energy.opDefColumn=operator definition
+Energy.powerPlatformColumn=associated power [W]
+Energy.objectiveLabel=Objective - in executions per second: 
+Energy.toleranceLabel=Tolerance - in %:
